requirements.txt
joblib==1.5.0
pandas
numpy
scikit-learn
matplotlib
scienceplots
seaborn
openpyxl
ipywidgets
ipython
pyyaml
wesanderson
plotly